feat: ego server stop|restart now use REST /service/admin/down endpoint
By default, attempting to stop the server will use a REST call to tell
the server to do a clean shutdown. This closes the log file cleanly,
etc.

Note that a ego server restart _cannot_ be done this way if the
server in the profile is not the current host -- you can stop a
remote server (if you have admin privilege) but you cannot start
a remote server, so restart is not allowed.

If you want/need to stop or restart the server on the local machine
without using the REST interface to stop the server, use the "--force"
flag which instead uses direct OS calls to kill the server based on
the PID number of the server. This does not require admin privilege
(but does require you to be the owner of the server process or have
root privilege).