feat: compile-time checking for references to unknown symbols
This is incomplete (dot-references are not checked, for example)
but most symbol references in the compiler are checked to verify
that they reference a defined symbol, package, etc. Additionally,
a symbol that is defined but not used generates a compile error.

Note that at this point, no type checking is done, only verifying
symbol existance.

This feature is controlled by the ego.compiler.unknown.var.error
configuration setting. This can be set to false if the feature
is found to be problematic.

This fix also involved updating several integration tests to conform
to the more stringent rules for symbol usage.