fix: use of global symbol defined in another source file not flagged as error
The compilation processor needs to defer unreferenced symbol errors until after
the compilation complete so there's an opportunity for the symbol to be
discovered to be a global in a different source file that might be compiled
after the current source file (multi-file project runs or packages).

The unresolved errors are accumulated in the compiler, and only when the
compiler is done being used (Close() called) does it validate that there
are unresolved symbols. When a global symbol is defined, if it was found
to be an unresolved reference, the unresolved reference is removed from
the compiler's deferred error list.

This fix also improves support for a clone of a compiler, such as when
a test expression is compiled with an instance of the current compiler
but might be discarded.  Close() will percolate state back to the
parent compiler if called, so do not call if an active error was already
caught by the compiler.