fix: try/catch will now correctly unwind call frames when error thrown
Previously, the error handler that checked for an active try block did
not consider that there might have been nested calls within the try
block before the error was enountered. The error handler now unwinds
the call stack by processing any call frames found on the stack to the
point where the try block was defined. This correctly resets the context
for execution to the state at the time of the try{} so the catch can
successfully resume.