feat: revised support for import packages
This is the initial push for updating packages. This work is ongoing and
not yet ready to merge with master.

1. Package caches are in a separate Go package "packages" that can be
   accessed by the compiler and the runtime.

2. Import scope is now set to the symbol table for the file that contains
   the import. No more storing all packages in the global symbol table.

3. Packages are now stored by path, not name. This allows user-created
   packages of the same name as an Ego package to both be used by the
   code. The cache is keyed on the path.

4. This is all a precursor to trying to massively simplify how package
   scope happens. Currently, a lot of symbol table movement happens when
   a table goes out of scope that was part of a package. Ideally this
   won't br required as a package will be a symbol table primarily.

Much of the incomplete work still revolves around managing package
symbol table scope correctly, so any source file that used import
statements gets the correct symbol scope added or dropped when
running functions within the package, etc.