fix: simplify @authenticated functionality
The scanner that collects up services searches the service definition
for @authenticated and accepts "none", "user, or "admin". These flags
are used to set the route authentication requirements for the service.
This removes the need for the running service to do authentication checks
as the router has already dealt with them.