feat: rewrite of HTTP request and response handling
This is the first major update to HTTP support. The RequestWriter
is now an Ego struct wrapper around the real ResponseWriter for
each request. It supports Write, WriteStatus, and Header calls.
The Ego wrapper also tracks how many bytes are written to the
response, so this can be extracted and tracked in the server
handler's summary log messages.

Ego services are being updated to honor the requirement that
the Write() service can only write a stream of bytes.

Additional work will be done to ensure that everything that was
formerly hidden in readonly global variables for the service
is now placed in the Request object, which will become a native
Ego object containing all the information from the session as
well as the actual request info.