fix: correct issues with receivers in source package files
There were a number of issues with how receiver values in source files
contributing to packages were handled that could result in an invalid
or incomplete receiver value.

Additionally, this addresses issues with symbol table lifetime when
package proxy symbol tables were added to the symbol stack during
"InPackage" processing, that could result in hoising the last
reciever value in to the package proxy and conflicting with the
"real" proxy on subsequent invocations.