fix: correct order issue with ResponseWriter writes vs headers
The Ego shim on Write was actually calling Write() for the body
of the response. However, the WriteHeader() call was only ever
made during end-of-service post processing. The Write() call was
flushing the headers prematurely before the status and headers
were retrieved from the Ego response writer object.

This change causes Write() to not call the native ResponseWriter
object, but instead buffer the bytes locally. They are retrieved
by end-of-service processing just like the status and headers,
so they can always be sent to the native Writer after all header
processing is complete.