feat: add $hash as dictionary value for apitest expressions
Using $hash as a value creates a unique (uuid-based) string of
characters, 25 characters long. This is shorter than a full
UUID (base 32-encoded) and makes for easier-to-read names
of objects. This is intended to create one-time name values
(by placing the values in the dictionary) for testing that
creates named objects that are guaranteed to not collide with
other concurrent uses of the server (including other tests)