fix: correct package builtin type definitions for self-referenced types
Package builtins often need to use their own types for defining a function
receiver variable type, a parameter, or a return value. This has to be done
as a fix-up operation since the static type definitions aren't allowed to
be recursive in Go.

Previously, this fixup operatino occurred as each function was being defined.
However, sometimes this was not correct since the function was bound to the
underlying structure type and not the package type. This caused the receivers,
etc. to all be labeled as having type "struct" instead of the actual type.

The fix is to defer the fixup until the type is otherwise complete, and
then run the fixup, with the fixup routine being sensitive to functions
at the level versus the underlying structure (or other) type.

Added code to the reflection integration tests to validate that the string
formatter for self-referenced type functions were coded correctly.