chore: remove apitest code and use tools/apitests to hold test files
The apitest is now a separate project (github.com/tucats/apitest) and
is run assuming it has been built and made available via the default
PATH.

The tools/ directory has been renamed apitests and now is the rool level
of the test directory structure. The tools/test.sh command has been
updated to use the CLI version of apitest and the new test directory.

A README file has been placed in the test directory to point to the
apitest tool and how to find doc on installing it.