feat: allow --sequence to specify sequence numbers or ranges for ego log command
The range can be a single range or a list of ranges separated by commas. If the
expression has spaces in it, it must be enclosed in quotation marks.

Each range is either a single value, or a start:end expression. If the colon is
present but the start is missing, 1 is assumed. If the end value is missing,
start+9 is assumed, producing a range of ten items. The end must not be smaller
than the start value. The numbers must be valid positive integer values.