feat: add supprot for (e *error) Context(), In(), and At()
This allows a user program to generate a more complete error value
with supplmeental context (string, location, and module).  Note
that the internal variables __line and __module can always be
referenced as the Ego code's current location if this is useful.

These values can be extracted from an error using reflect.Reflect()
on the error, which has a Context result value containing the
three context variables.