feat: implement first router-based validation tests
A route can how have a list of validation suites. If present, at
least one of the validation suites must pass with the payload
provided by the REST client request. If non pass, the validation
error is returned to the caller automatically, before the service
is even invoked.

Added a new apitest to verify that the valication works (by passing
an incorrect payload for a user create operation).