feat: make ego memory log interval a configuration item
ego.server.memory.log.interval is a duration, and is the interval
between checks to see if a service has been called and if it's time
to generate a memory log entry. If not present (the default) it assumes
a 5 minute interval.

If the interval is invalid, it is reported once (at server startup)
and the falls back to the default of 5m.