feat: allow aliases to rest payload validations
By default, the router will assign a validation if it exists in the
validation dictionary, by forming a name from the endpoint path (including
any parameters) and the method. If this was found when validations were
loaded, it becomes the default validation for the route (this can be
overridden by using .Validations() method on the route).

Because multiple paths might use the same payload format (/admin/users POST
and /admin/users/{{name}} PATH for example) there can be a single type
definition ("@users" in this case) that is referenced by multiple path
specifications.

Additionally, all validations are now read from the lib/validations directory
and can be organized arbitrarily by functionaliry, etc. in different .json
files. For example, loggers.json contains the validations for the /admin/logger
endpoints, and users.json contains both /admin/users endpoints and the credentials
validation used for logon operations, etc.