feat: some panics don't need a full stack trace, so offer logged panic
IF EGO_PANIC is set to a value that means "true", then a real panic
is done anyone. Otherwise, the message is logged to the SERVER logger
if present, else the INTERNAL logger if present, else printed to the
console and then the program stops with a status value of 99.