feat: improved /admin/validation parameter support
You can specify a path and method and Ego will look up the validation(s)
for that entry point and return them.

   method=PUT|PATCH|POST|DELETE|UPDATE|GET
   path=/url-path

If a route exists for this method/url combination, then the validation(s)
if any are returend for it.  Alternatively, you can request a specific
entry to be returned using

   entry=ditionary-entry-name

The parameter is the specific name of the requested validation.  You cannot
specify ?entry along with method or path.  If neither entry or path
parameters are used, then the entire dictionary is returned.