feat: add ego server validations command to fetch payload validation
You can specify --all to get all the validation rules at once.

You can specify --entry to request a specific entry and any of its
dependent definitions.

You can specify --path to give a URL path with optional --method
specification (default method is POST) to look up the corresponding
validation(s) for that endpoint.

This fix adds an additional feature to the CLI grammar processor where
options can be expressed as mutually exclusive, so the validation is
done during CLI processing and the subcommand handler does not have
to do so.

Also fixed a bug in the tests/packages/uuid.ego test caused by the
previous correction of the Gibberish() method.