feat: allow route spec to include disallowed parameter combos
When defining a route, you an specify sets of disallowed combinations
of URL parameters not permitted by a given handler. This validation is
done as part of route selection before the handler is run. This means
the individual handlers are guaranteed not to have invalid parameters
present in a given request, and eliminating the need for per-handler
code to check for this.