feat: start support for reflection-based validation definitions
This allows a "valid" tag to be added to structures, so when an
instance of the structure is passed to validation.Reflect(), a
new dictionary entry is created for the sturcture and it's nested
components.

This is still under test -- the definitions being used are all
still currently read from the lib/validations/*.json files. But
this will help with more easily creating definitions for internal
structure types, leaving the json files there for defining payloads
for Ego rest services generated by the user.