feat: move payload validation "inboard" for DSNS, user info
The lib/validations/dsns.json and users.json files have been replaced
with "valid" tags on the internal structures. The validation initializer
use the validation package to create definitions for structures using
the "valid" tag.

Note that loggers.json still exists, as it contains a mapping for
the contents of a map object in the structure, which is not yet supported
by the refection-based validation definition method.