feat: fix issues with array element validations
The validations were being duplicated on the field in the object
that points to the dependent array type, and also in the array
type. This fixes them so they only exist in the array item definition.

Changed array size validation items to "minsize" and "maxsize" so they
are different that "min" and "max" which refer to elements in the
array. Also added "minlen" and "maxlen" for string length constraints.

Added new docs/VALIDATION.md to describe the validation mechanism.