fix: correct validation initializaiton errors regarding array names
The tag value was not being properly processed at the structure field
level for an array, resuling in the name being the default (field name)
when it is probably overridded by a json or valid tag name.

Also added VALID logger and moved all logging messages related to
payload validation to this logger, so they can be easily isolated
when debugging, but turned off otherwise.