feat: Global --json-query query expression for JSON output
For any command that produces JSON output when --format json or
--format indented is used, you can instead specify a JSON query
string with --json-query "<query-string>". This defaults the
output format to JSON, but then uses the query string to parse
the resulting JSON to pull out specific elements. See the doc
for https://github.com/tucats/apitest for a summary of the
query string format.