feat: Enhanced json query expressions allowing index of "all elements" result
For example, you can query the server cache results json from the ego server cache
command with:

   ego -j items.*.name.1

This looks for the element items, reads all array elements, searches for the
field "name" in each array element, and returns the second (zero-based) memeber
of that array of results.

In the above example, since all array elements in this JSON payload are
homogenious, the same result could be gotten with:

   ego -j items.1.name

However the use of the new feature (indexing into an all-elements result)
allows looking for members of non-homogenious object types.