fix: some compiler resulted in nil error incorrectly reported
When an error was returned that was a Ego error but was a nil error,
it was treated as a non-nil error and attempted to be reported. This
showed up when ego.compiler.unknown.var.error was false and an unknown
variable was used. This returned an empty Ego error value but was a non-
nil pointer, so the REPL loop attempted to report it as an error, when
there was no error.

Also fixed a missing localization for the ego config describe output
for ego.compiler.unknonw.var.error