fix: prevent unnecessary duplicate decryption of token on rest authentication test
The code was first validating the token, then extracting the username to look up
permissions. However, the name extraction called the validation function again.
The fix is to just attempt to extract the name from the token; if there is an
error report it, else use the username and continue.