chore: using new apitest feature, remove payload files and use "body" for request info
The updated apitest tool allows inline "body" expressions for the payload to use
with the request when it is JSON. The update allows the body to contain replacement
strings of the format {{name}} where the value "name" is replaced with whatever the
active dictionary has for that named value. This allows saving payload responses for
later use. Now that apitest handles these substititons properly in the JSON test
description, we no longer need separate .payload files that are processed for
dictionary value.