fix: ego table insert command needs to coerce parameters based on column types
The insert command needs to correctly coerce the text representation of the
data item from the command line to match the database column type. This allows,
for example, expressing a string value as a long string of digits without the
cli parser converting it to a floating point numeric value along the way.

This does require the ego table insert command to make two REST API calls,
one to get the column metadata for the given table, and one to do the actual
insert.