feat: add upsert ability to table row PUT endpoint
If the parameter ?upsert (or ?upsert=true) is added to the endpoint,
the insert becomes an upsert operation. For any row in the data that
has a _row_id_ value already, that row is updated, not inserted.
This allows for the intermixing of inserts and updates based on
whether there is already a non-empty row id value found in each
row of the rowset payload on the request.

If the upsert flag is not specified, any _row_id_ value is ignored
and all rows are inserted with a new unique row id value.

Also, updated the util.Log() function to return a log entry value
indicating there is no active log, rather than an empty log set
when there is no log. This allows the apitests to run against a
server not being run as a detached process (such as debugging).