feat: allow ?upsert parameter on table PUT to specify column names
By default, ?upsert alone will use the row id value as the key to
determine if a given row in the payload is to be updated versus
inserted. If you specify column names as the upsert parameter values,
then those are the columns used to determine if the operation is an
update versus insert. This is determined by detecting if one or more
rows already exist in the database with the given column containing
the value from the payload. If so, an update is done, else an insert.

This change also updates the API doc to specify this parameter.