feat: automatic row id values is configuration by dsn
A dsn by default has this set to true, but can be overridden using
the ego dsn add --row-id=false option. When set to false, tables
created with this dsn do not automatically have an extra column
added that contains a unique row id value. Note that this attribute
applies to the specific DSN used to access the table.

The column name "_row_id_" is still a reserved name and cannot be
specified when a table is being created.

Note that using the DSN-less access mode (endpoint /tables) will
still always create a table with a unique _row_id_ value. This form
of access is deprecated.