feat: alternate cli grammar when EGO_GRAMMER=verb
The alternate grammar is more verb/direct-object, so instead of

   ego server restart

the alternate form is

   ego restart server

Note that the altnernate grammar is incomplete at this point. Additionally,
when starting up, ego will always look in ~/.ego/env.json for a JSON object
that contains environment variable names and values, and will set those
variables in the EGO process. So you can put environment variables here that
will always be set when the program runs, without you having to put them
somewhere else (like a shell .rc file, etc.)

IF the env.json file does not exist, is not readable, or contains invalid
JSON, then no error is reported and the environment variables are not set.

Non-empty environment variables set in the shell when running Ego will
not be replaced by the env.json variables; the shell values take
precident.