fix: do not manage admin.permissions when using a DSN-connected database
the admin.permissions table should only be used when the serve-wide "default"
database is accessed via the /tables endpoint (this is being deprecated). If
the connection is done via the /dsns endpoint, no permissions checking is done
at the Ego server level, as the connection has passed credentails to the
underlying database and it is responsible for validating permissions.