feat: use json validation to verify env.json file contents
The env.json file, located in the ".ego" profile directory, can
contain environment variable definitions that are set before the
Ego program starts running. For example, it can be used to set the
EGO_GRAMMAR variable to either "verb" or "class" depending on the
style of grammar definition preferred by the user.

This fix applies the JSON validator to the contents of this file.
For example, if EGO_GRAMMAR is set to a valud other than "verb"
or "class", an error is generated when Ego starts.

The validation is controlled by the "lib/validations/env.json" file,
sos validations can be added or modified as needed, including allowing
additional user-specified environment variable names.