feat: allow user-defined env.json definitions
The user can add application- or user-specific values to the .ego/env.json
file to have values automatically added to the environment of the running
Ego program. Note that these names cannot start with "EGO_" or "ego." which
are reserved to Ego

This fix includes adding the "foreign": true key to the validation json
in lib/validatsion/env.json which allows for undefined key values to be
included in the JSON. It will still catch mispelled EGO_ prefixed names,
but allows arbitrary environment values without that prefix to be defined.

Also added LookupEnv and ExpandEnv functions to the os package to assist
with using environment variables.