fix: cleaner and better chaining of error messages
With this fix, any error Chain() operation causes the next item to
be chained to always be added to the existing stack at the end, and
the message(s) are printed in that order.

This fix eliminates the errors.Chain() method whose use somewhat
inconsistently allowed arbitary construction of chains (prepend
vs append).

This also results in somewhat cleaner messages geenrated from the
JSON validation module. Unit and API tests updated accordingly.