feat: add os.Expand(), os.Setenv() functions
The os.Expand function will expand values in a string that use "$"
subsituttion operators, similar to ExpandEnv(). The caller must
provide a lookup function as the second parameter that either returns
a substitution string value or an empty string.

Calling os.ExpandVar("test") is a simplification of calling the
function os.Expand("test", os.Getenv) since the os.Getenv()
function has the required function signature to be a lookup
function.

This fix adds the ability to specify a Go-native runtiem function
as a parameter to another function if needed (such as the os.Getenv
parameter to os.Expand in the above example).

This fix also adds a native os.Setenv() function which will set an
environment value in the current process environment space.

This fix adds a unit test to test os.Expand with an Ego lookup
function, a builtin runtime function, and the os.Setenv native
function.