feat: incoming bearer tokens can be checked against a blacklist
If a token string gets "out in the wild" it can be added to a blacklist
table in the credentials database. If the token is found on the blacklist,
it is logged to the AUTH logger as a blacklist, but reported to the end
user as an invalid token encryption.

The blacklist is stored in the table "blacklist" in the credentials database.
This is not supported when using the in-memory credentials database mode.

Future feature will allow for an API to add items to the blacklist instead
of doing it manually.