feat: adopt new validator package for handling env.json validation
The externalized JSON version of the valudator is found in
lib/validations/env.json and matches the format of the newly
adopted github.com/tucats/validator package. A minimal README
file was added describing the layout of the validator. This
allows an end-user to add validators to the env.json file for
user-defined environment variable settings.

Note this directory can also be used to extend validations on
REST payloads the same way, by creating files that contain
dictionary entries for the validator, where the key is the name
of the endpoint followed by a colon and the method tag.