fix: remove external validator definition for env.json
At the time during startup that the env.json file is read (and
validated), the configuration subsystem cannot have initialized
yet and there is no way to know the correct patht to the /lib
directory where the validation file should be found.

So, in this update, the validation is always statically defined,
in the app-cli/app/app.go file, which has a structure that contains
all the known (and validatable) env.json entries. The file can also
accept user-defined entries which are defined in the process
environment, but no validation is done for them.

In place of the env config validator, a sample validator for
an endpoint /sample/user for a POST operation is provided as
an example validation that can be created by the user for user-
created REST endpoints.