fix: quoting for --filter parameters on table CLI commands corrected
There were cases where a value was given in a filter expression, such
as a UUID literal, that would be re-parsed on the server side as multiple
tokens. This fix cause any such expression values in the filter to be
quoted on the client side first.

Also, corrected an issue with prompt strings for the CLI not being
washed through the i18n Label conversion.