fix: fully resolve ego executable name before starting server
This solves some ambiguous file name issues. The updated code uses
the PATH search to see if it can be found. If not, it uses the
fully qualified name as given. In either case, the full name is
used as the image name when starting the server, and is what is
stored in the PID file.