fix: _buildtime global variable should be of type time.Time, not string
IF for some reason the build script does not inject a valid buildtime
string into the app, _buildtime will just be a copy of that string value.
However, if it can be parsed as a time value (regardless of format or
layout) then _buildtime will be a time.Time value instead.