fix: limit rest request headers written to log
The log had been dumping all but the authentication headers to
the log. However, there are plenty of other headers associated
with security and routing management that shouldn't be leaked
as readable text.

The fix is to allow a specific set of headers (relevant to Ego
processing) to be printed to the log, but no others.