feat: If -v specified on server start or restart, print more info
If -v is specified and output is in text mode, the output will
include the server UUID, the invoking impage path, and the log
file name (sans the date stamp added to the end of the name).

If -v is omitted, only the server pid number is printed. If the
output format in JSON, the -v is irrelevant as the entire pid file
contents are send to stdout as JSON, which includes all the above
information.

Also, if not in text output mode, restart does not display chatty
messages about stopping the server, only displaying the JSON in
the event of a successful restart.