fix: refine how help is displayed for incompelte commands
By default, the message still shows the list of expected subcommands
and a note about using --help to learn more.

If ego.console.auto.help is true, then instead of the message about
using "--help", the acctual help text is displayed. In this case, the
copyright header is not displayed as it breaks up the flow of info
from the help text.

In all other cases, the help information includes the copyright
information.