feat: simplified localizataion tag "card" for magnitude strings
Instead of using:

  {{count|zero "no frogs"|one "frog"|many "frogs"}}

You can now use the simpler/shorter version:

  {{count|card "no frogs","frog", "frogs"}}

The cardinality strings are for the value being zero, being one,
or being greater than one, just like the zero, one, and many
magnitude tags.

An even shorter form can be given when the magnitude strings are
for one versus not-one, using just two values:

   There {{count|card is,are}} {{count}} {{count|card row,rows}}

When the card operator (or it's longer form, cardinality) is followed
by only two values, they are assumed to be for one and not-one. So
the above message text will output:

   There is 1 row
   There are 0 rows
   There are 9 rows

The count value is used three times, for the cardinality of the
verb, for the actual number reported, and for the cardinality of
the suffix label.

The inteent is to allow localization strings that express plurality
and cardinality without requiring a separate message for each case,
letting the formatter determine which term(s) to select based on a
numeric value.