feat: add strings.Substiution() which uses Ego JSON query expressions for substitutions
This is the same basic functionality as the egostrings substitution funcions (which it
calls) but first it converts the arbitrary value(s) passed as the second (or more)
arguments into a map that can be queried. The map construction is based on using the
JSON query expression syntax to identify items in the passed parameter.

For example, a query expression of {{User.Name}} could find a structure member User
and within that a structure member or map value for a key "Name".