feat: defaults.json in lib directory contains default config settings
Whenever a new profile is created, these defaults (if found) will be
read in and stored in the profile that is newly created. These can
override the default values stored by the Ego language processor, or
add any other configuration values desired by the administrator. These
values can of course still be changed by the user at will -- this is
just a template of configuraiton items that can be used to give each
profile the same starting place.