feat: add util.Package() which returns map[string]map[string]string describing elements
The top level map has up to four members, "types", "functions", "constants", or "variables".
Within each sub map the key is the name of the item, and the string value is text expression
of the type, constant, variable, or function. This allows programatic navigation of a package
structure if desired.