chore: refactor settings persistence to interface for future expansion
In the future, settings can be read from a database rather than the
file system. This will support using Ego in a container or other environment
without persistent local storage. The default is to use the file system as
has previously been done. If the environment variable ENV_CONFIG exists
it is a URL to the storage area for the configuration. Currently, only
file:// is supported, but in the future sqlite or postgres will be valid
URL schemes as well, and will connect to the given database.