feat: ability to read configuration from external database source
If the EGO_CONFIG environment variable is non-empty, it contains a
URL to the configuration data. This can be a file:// URL that points
to the name of the configuration object to read, which uses the existing
file-based configuration system.

It can also be a sqlite3://path/to/data.db expression which uses a
Sqlite3 database to contain the informaiton.

Finally, it can be a postgresql://conection-string URL which accesses
the named database, and reads from config_ids and config_items tables
in that database location.

When a database is used to store the configuration, config changes are
written back to the database if the database permissions allow it.

The intent of the remote configuration is specifically to support running
Ego in a container or other area without persistent storage where either
initial configuration or ongoing configuration changes may be made.

Note that when a databse config method is used, when creating a new profile
that does not already exist, the profile is loaded with any values found
in the config_items table with a UUID with the nil UUID value. That is,
these become the initial defaults for any new configuraiton created.