chore: refactor configuratio persistence services further
Refactor the persistence service objects so they can be used outside
the direct active configuration used by the settings package. This
means removing any references to the current configuration data from
the services, and providing simplified wrapper functions for the default
use by the settings package.

This is in anticipation of offering a config service API that can allow
an Ego server to manage it's configuratio remotely via another Ego server.