feat: support SIGINT when REST server running
A SIGINT sent to a running rest server will shut it down, as if a server DOWN
api call had been made. This also allows for neater cleanup when debugging a
server from CLI mode.

Additionally, cleaned up the SIGINT trap in bytecode execution so the signal
is only ever handled once, and the signal trap is disabled whenever the Run()
method executes, so multiple Go reoutines are not left around after each
bytecode execution.