feat: promopt for missing required command line options
IF the config item ego.console.prompt.missing.options is "true" then
the CLI will prompt for missing but required option values. For example,
creatinga DSN but not specifying the database provider type will result
in a prompt for the provider type on the console.

If the configuration is not "true" then the missing value just results in
an error.