feat: add ego.server.default.log.file configuration item
This allows specifying a relative or fully-qualified file name that
is used as the default log file when starting a server. This default
can be overridden by using the --log-file= command line option or
setting the EGO_LOG_FILE environment variable.

By setting this default, the Ego server can be started from any
working directory, and the log file won't by default be placed in
the current directory, but instead in the designated location.