feat: better handling of default database name for first server start
If the system database is not explicitly defined in the configuration,
it now defaults to "ego-system.db" as a sqlite3 database in the default
ego path location.

Additional fixes to ensure that the default name is handled correctly
when starting and restarting a server. The ego.server.userdata config
item can override the default name (and can be a complete database
URL, or a file system reference).

Also added new table to the database instance of the authorization
for a table called "starts" which tracks the start time of server
instances.