feat: add --query to log formatter to allow arbitrary JSON queries for log filtering
If --query (or -q) is present, it should contain an exprssion using JAXON query
syntax to specify elements of the JSON for each line to determine if the line
should be printed by the formatter. The expressions can either test for existence
by using a simple JAXON query, or can use operators to test for equality or
rlationship between numeric values.  Operators supported are
   "=", "!=", "<", "<=", ">", ">="

So, for example, to find log entries with a length field greater than 100, use:

   -q "args.length>100"

To search for a speecific string value, use single or double quotes:

   -q "args.key='Accept'"

This query can search for log entries with a header key euqal to "Accept".

Currently, compound query strings are not supported.