feat: if media type is text for /services/admin/log, log is formatted as text on server side
This allows a client to request the log as human-readable text. The payload is returned
with media type of text/plain. Note that the formatting of the log JSON entries is done
on the server side in this case, and so localizations will always be in the locale of
the server.

This differs from reading the log as JSON which sends locale-independent logging information
which can be converted locally as needed (see the ego log formatting commands for examples).